#163 remove core:date module
