+++ core/common/kotlin/src/main/kotlin/com/moobeside/core/common/kotlin/extensions/DateExtension.kt
@@ -1,10 +1,26 @@
-package com.jslee.core.date
+package com.moobeside.core.common.kotlin.extensions
+
+import java.text.SimpleDateFormat
+import java.util.Locale
 
 /**
  * MooBeside
  * @author jaesung
  * @created 2024/01/14
  */
+
+fun String?.transformDate(from: DateFormat, to: DateFormat): String {
+    if (this.isNullOrBlank()) return "알수없음"
+
+    val inputFormat = SimpleDateFormat(DateFormat.getFormat(from), Locale.getDefault())
+    val outputFormat = SimpleDateFormat(DateFormat.getFormat(to), Locale.getDefault())
+
+    val date = runCatching {
+        inputFormat.parse(this)
+    }.getOrNull() ?: return "알수없음"
+
+    return outputFormat.format(date)
+}
 
 enum class DateFormat(val pattern: String) {
     YEAR_MONTH_DAY_HYPHEN("yyyy-MM-dd"),
@@ -15,7 +31,7 @@ enum class DateFormat(val pattern: String) {
 
     companion object {
         fun getFormat(dateFormat: DateFormat): String {
-            return when(dateFormat) {
+            return when (dateFormat) {
                 YEAR_MONTH_DAY_HYPHEN -> YEAR_MONTH_DAY_HYPHEN.pattern
                 YEAR_MONTH_DAY_SEC -> YEAR_MONTH_DAY_SEC.pattern
                 YEAR_MONTH_DAY_MILLIS -> YEAR_MONTH_DAY_MILLIS.pattern
