+++ core/ui/build.gradle.kts
@@ -35,7 +35,7 @@ android {
 
 dependencies {
     implementation(project(":core:designsystem"))
-    implementation(project(":core:date"))
+    implementation(project(":core:common:kotlin"))
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
