+++ presentation/build.gradle.kts
@@ -47,9 +47,9 @@ dependencies {
     implementation(project(":core:ui"))
     implementation(project(":core:external"))
     implementation(project(":core:deeplink"))
-    implementation(project(":core:date"))
     implementation(project(":core:logger"))
     implementation(project(":core:image"))
+    implementation(project(":core:common:kotlin"))
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
