+++ presentation/src/main/java/com/jslee/presentation/feature/castdetail/model/CastDetailUiModel.kt
@@ -1,10 +1,10 @@
 package com.jslee.presentation.feature.castdetail.model
 
-import com.jslee.core.date.DateFormat
-import com.jslee.core.date.transformDate
 import com.jslee.core.ui.extension.roundVoteRate
 import com.jslee.domain.model.Gender
 import com.jslee.domain.model.Person
+import com.moobeside.core.common.kotlin.extensions.DateFormat
+import com.moobeside.core.common.kotlin.extensions.transformDate
 
 /**
  * MooBeside
