+++ settings.gradle.kts
@@ -29,7 +29,6 @@ include(":core:designsystem")
 include(":core:ui")
 include(":core:external")
 include(":core:deeplink")
-include(":core:date")
 include(":core:logger")
 include(":core:image")
 include(":feature:home")
